+++ src/controllers/ProofsController.ts
@@ -3,6 +3,7 @@ import express, { Request, Response } from 'express';
 import Block from '../models/Block';
 import Leaf from '../models/Leaf';
 import { AuthUtils } from '../services/AuthUtils';
+import { BlockStatus } from '../types/BlockStatuses';
 
 @injectable()
 class ProofsController {
@@ -20,10 +21,10 @@ class ProofsController {
       return;
     }
 
-    const block = await Block.findOne({ status: 'finalized' }).sort({ height: -1 }).limit(1);
+    const block = await Block.findOne({ status: BlockStatus.Finalized }).sort({ height: -1 }).limit(1);
 
     if (block) {
-      const keys = request.query.keys as string[];
+      const keys = (request.query.keys || []) as string[];
       const leaves = await Leaf.find({ blockId: block.id, key: { $in: keys } });
       response.send({ data: { block: block, keys: keys, leaves: leaves } });
     } else {